pyproject.toml
[tool.poetry]
name = "jumpstart"
version = "0.1.0"
description = ""
authors = []
license = "MIT"
#include = ["jumpstart/pkg-index/*","jumpstart/templates/*" ]
#packages = [{ include = "schema_graph" }]

packages = [
    { include = "jumpstart" },
]
# include = ["",]
# exclude = ["my_package/excluded.py"]

[tool.poetry.urls]
"Github Repository" = "https://github.com/imochoa/jumpstart"

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.6.0"
rich-cli = "^1.8.0"
typer = "^0.6.1"
marshmallow = "^3.17.0"
marshmallow-dataclass = "^8.5.8"
cogapp = "^3.3.0"

#typing-inspect = "^0.7.1, !=0.8.0"
# 0.8.0 was problematic

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
pytest = "^7.1.2"
pytest-icdiff = "^0.5"
pytest-xdist = "^2.5.0"
pytest-sugar = "^0.9.5"
docker = "^5.0.3"

[tool.poetry.scripts]
jumpstart = 'jumpstart.console:run'

#[tool.poetry.plugins."blogtool.parsers"]
#".rst" = "some_module:SomeClass"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
multi_line_output = 3
import_heading_stdlib = "stdlib imports"
import_heading_firstparty = "1st party imports"
import_heading_localfolder = "local imports"
import_heading_thirdparty = "3rd party imports"
force_sort_within_sections = true

[tool.black]
line-length = 121

[tool.mypy]
python_version = "3.10"
exclude = [
    'venv',
    '.venv',
    'tests',
    'bin',
]
# marshmallow_dataclass.mypy -> Required so that mypy recognizes the custom marshmallow NewType's
plugins = [
]
warn_unused_configs = true
warn_unused_ignores = true
show_error_codes = true
strict = true
implicit_reexport = true
warn_return_any = false
allow_untyped_defs = true

[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py"]
addopts = "--failed-first --showlocals -ra"
# Run in parallel ->  --numprocesses auto

[tool.bandit]
exclude_dirs = ["venv", ".venv", "tests"]
skips = ["B101", "B603", "B404", "B108", "B110", "B311"]

# vim: ft=dosini
